# DDP: 10% of NCCL backend perf improvements with mixed-prec support