This PR lets NCCL backend to directly enqueue the NCCL reduction kernels in the same thread since it's an async call by nature, and everything now goes to the default stream. This essentially gets rid of the overheads of python thread synchronization, stream synchronization, as well as bucketing map lookup overheads.  

For DDP with multiple GPU (the default use case), instead of doing a two step reduction, we use the new NCCL backend API to all-reduce all GPU at one time, this is basically 4 times all-reduce throughput on DGX1s compared to the all-reduce(single GPU version) plus the intra-node reduce overheads.

As a results, we have the following perf improvements.

For DDP with 8 GPU (default use case):

On two DGX1s with 8 V100s, 256 batch size / process, single process / node
ResNet50
**0.139 sec / iter** reduced from **0.154 sec / iter** (about 10 percent improvements)
ResNet101
**0.247 sec / iter** reduced from **0.272 sec / iter** (about 10 percent improvements)

For DDP with 1 GPU (multi-process use case)

On singe DGX1s with 8 V100s, ResNet50, 32 batch size per GPU and process, 8 processes distributed training 
**0.109 sec / iter** reduced from **0.116 sec / iter** (about 6 percent improvement)

In addition,  added bucketing to limit the memory usage, added mixed precision support.