@apaszke 2018-02-13
I'm sorry, why should it be the same? Ethernet is likely to have much lower throughput than 4xIB that you tested with, and interleaving the communication with backward might still be beneficial in that case.