@teng-li 2018-02-13
@apaszke I meant, we don't currently support multi-threading for NCCL. Remember that the gradient buckets of each process on different nodes need to be executed in the exact same order. So even if we use the old code path, we still limit the number of bucket to be 1.

I will further test the Ethernet perf, which is on my to-do list anyway.  But this should block this PR for now