@apaszke 2018-02-14
Is this true, even if they are using different groups? I thought that it doesn't matter as long as they use different communicators, right?