@teng-li 2018-02-14
@apaszke right, they are using different communicators.  But the NCCL call order (among all the reduction thread) needs to be maintained among all the nodes. The gradients available order for each bucket can not be guaranteed for different nodes.