@apaszke 2018-02-15
Does it have to be the same, even when they are on different communicators? I think the whole purpose of keeping multiple comms was to allow concurrent operations to execute independently