@csarofeen 2018-02-15
I believe if your communicators overlap, you can get in trouble if the order is not correct. I don't believe you can have 2 communicators on 2 gpu's and have:
GPU 0 call on comm 0 then comm 1
and
GPU 1 call on comm 1 then comm 0