@apaszke 2018-02-16
Aren't these situations the reason why `ncclGroupStart` and `ncclGroupEnd` were introduced?