@teng-li 2018-02-16
@apaszke  comments addressed