@ngimel 2018-02-16
Got answer from nccl: "calling collectives on different comms is to be avoided, and if there is no other solution it should be properly ordered". Apparently, the purpose of different comms was not to use them.