@teng-li 2018-02-20
@apaszke  The latest commit has been tested with ResNet50 with the correct 76 accuracy on two nodes. It should be safe to land