@apaszke 2018-02-21
@teng-li can you please fix the conflict?